Added a check for declarative-private module to enable building on Ubuntu 14.04 with system QT library.
I'm not sure why, but even when the declarative-private package is installed on Ubuntu 14.04, it still is not found by qmake (at least on my machine).
Adding this line (which was also added to the qtcreator source in commit I1a58d9bce7c0c6409b0473e81913d7d207b88444.

Signed-off-by: Ted Steiner <tsteiner2@gmail.com>